+++ application/src/main/java/org/thingsboard/server/install/ThingsboardInstallService.java
@@ -170,6 +170,7 @@ public class ThingsboardInstallService {
                 log.info("Installing DataBase schema for entities...");
 
                 entityDatabaseSchemaService.createDatabaseSchema();
+                entityDatabaseSchemaService.createSchemaVersion();
 
                 entityDatabaseSchemaService.createOrUpdateViewsAndFunctions();
                 entityDatabaseSchemaService.createOrUpdateDeviceInfoView(persistToTelemetry);
