+++ application/src/main/java/org/thingsboard/server/service/install/EntityDatabaseSchemaService.java
@@ -23,4 +23,6 @@ public interface EntityDatabaseSchemaService extends DatabaseSchemaService {
 
     void createCustomerTitleUniqueConstraintIfNotExists();
 
+    void createSchemaVersion();
+
 }
